Socket mode for buffer sidecar
